requirements.txt
flask == 2.3.3
python-dotenv == 1.0.0
flask-SQLAlchemy == 3.0.5
flask-marshmallow == 0.15.0
flask-Cors == 4.0.0
marshmallow-sqlalchemy == 0.29.0
sqlalchemy-filters == 0.2.0
psycopg2 == 2.9.7
Flask-Migrate == 4.0.4
psycopg[binary]
